cli: make input file a positional argument
This patch slightly modifies the command line parser in order to
simplify the syntax and allow the user to provide the file path as a
positional argument. Since gumdrop doesn't seem to allow both free and
command arguments at the same time, the subcommands ('at', 'stats' and
'when') are now exposed as optional options.

Before:

    $ dwfv -i file.vcd
    $ dwfv -i file.vcd stats
    $ dwfv -i file.vcd at 42

After:

    $ dwfv file.vcd
    $ dwfv file.vcd --stats
    $ dwfv file.vcd --at 42

Signed-off-by: Pierre Surply <pierre.surply@gmail.com>