Allow the crate to be used as a library
This adds a default feature flag that can be disabled to allow dependents
to use the crate as a library without pulling in the transitive
dependencies that are only used for the TUI.

This is especially useful on Windows, because `termion` does not build
at all on the platform.

Tests pass on Windows when running with `cargo test --no-default-features`.